+++ ApplicationCore/PortfolioAggregate/IPortfolioService.cs
@@ -5,7 +5,7 @@ namespace ApplicationCore.PortfolioAggregate
 {
     public interface IPortfolioService
     {
-        public Portfolio CreatePortfolio(int userId, string name, double initialCash, string initialCurrency);
+        public Portfolio CreatePortfolio(int userId, string name, decimal initialCash, string initialCurrency);
         public List<Portfolio> GetPortfolioList(int userId); 
     }
 }
