+++ PublicAPI/Endpoints/Portfolio/CreatePortfolioRequest.cs
@@ -3,7 +3,7 @@ namespace PublicAPI.Endpoints.Portfolio
     public class CreatePortfolioRequest
     {
         public string Name { get; set; }
-        public double InitialCash { get; set; }
+        public decimal InitialCash { get; set; }
         public string InitialCurrency { get; set; }
     }
 }
